requirements.txt
# HCDS 依赖项
# Hierarchical Clustering + Dynamic Sampling

# ==================== 核心依赖 ====================
numpy>=1.21.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.60.0
psutil>=5.8.0

# ==================== 嵌入计算 ====================
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# ==================== 聚类 ====================
hdbscan>=0.8.29
scikit-learn>=1.0.0

# ==================== 向量检索 (大规模) ====================
faiss-cpu>=1.7.0
# 或者 GPU 版本: faiss-gpu>=1.7.0

# ==================== 数据存储 ====================
h5py>=3.7.0
pyarrow>=10.0.0  # Parquet 支持

# ==================== 可视化 (可选) ====================
matplotlib>=3.5.0

# ==================== HuggingFace 数据集 (可选) ====================
datasets>=2.10.0

# ==================== OpenAI API (可选) ====================
openai>=1.0.0

# ==================== 分布式 (可选) ====================
# ray>=2.0.0

# ==================== 开发依赖 ====================
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# isort>=5.10.0
# mypy>=1.0.0
